Fix adding items to rule build out
When we add an item to the "rule" that we're building out in the
widget, we need to make sure that we're actually adding the item
that the user clicked on. The on-screen representation of a rule
at different teirs is actually a complex rule, so we must first
subtract the predicate items of the prior rule such that we can
isolate the item that is actually being added.

This worked before but became broken with the r22 updates to the
API endpoints.